[Java] P ints should be extracted as Java Longs (#483)
* [Java] P ints should be extracted as Java Longs

Other backends use the widest integral types available on the system.

* [Java] downcast seq accesses to an int

With the change that P int types are now longs, a special-case
cast is required to maintain the Java container API.

+++ Src/PRuntimes/PJavaRuntime/src/test/java/testmonitors/clientserver/ClientServerTest.java
@@ -19,8 +19,8 @@ import static testmonitors.clientserver.PTypes.*;
 
 public class ClientServerTest {
     private BankBalanceIsAlwaysCorrect initedBankBalanceIsAlwaysCorrect() {
-        HashMap<Integer, Integer> initialBalances =
-                new HashMap<>(Map.of(100, 42, 101, 10));
+        HashMap<Long, Long> initialBalances =
+                new HashMap<>(Map.of(100L, 42L, 101L, 10L));
 
         BankBalanceIsAlwaysCorrect m = new BankBalanceIsAlwaysCorrect();
         m.ready();
